+++ src/forex/oanda/headers.py
@@ -1,8 +1,9 @@
+from forex.oanda.config import oanda_config
 
-def get_oanda_headers(config):
-    headers = {
+
+def get_oanda_headers() -> dict:
+    return {
         'Content-Type' : 'application/json',
-        'Authorization' : 'Bearer ' + config['token'],
-        'Accept-Datetime-Format' : config['oanda_date_time_format'],
+        'Authorization' : 'Bearer ' + oanda_config.OANDA_TOKEN,
+        'Accept-Datetime-Format' : oanda_config.OANDA_DATE_TIME_FORMAT,
     }
-    return headers